Move timeout tests
